unified project for SkypeWeb protocol
git-svn-id: http://svn.miranda-ng.org/main/trunk@13310 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/SkypeWeb/src/http_request.h
@@ -253,7 +253,7 @@ public:
 
 		if (!pData) {
 			pData = Body.ToString();
-			dataLength = mir_strlen(pData);
+			dataLength = (int)mir_strlen(pData);
 		}
 
 		char message[1024];
